Q&D slab implementation

+++ include/firefly/memory-manager/mm.hpp
@@ -12,10 +12,10 @@ enum AddressLayout : uint64_t {
     // clang-format on
 };
 
-enum PageSize : uint64_t {
-	Size4K = 0x00001000,
-	Size2M = 0x00200000,
-	Size1G = 0x40000000,
+enum PageSize : uint32_t {
+    Size4K = 0x00001000,
+    Size2M = 0x00200000,
+    Size1G = 0x40000000,
 };
 
 static constexpr uint32_t PAGE_SHIFT = 12;  // Lower 12 bits of a virtual address denote the offset in the page frame
